refactor: refactored according to pr
