+++ src/common/constants.ts
@@ -7,3 +7,11 @@ export enum Services {
   CONFIG = 'IConfig',
   OBJECT_STORAGE = 'IObjectStorage',
 }
+
+export const NAME_LENGTH_LIMIT = 100;
+
+export const BUCKET_NAME_LENGTH_LIMIT = 63;
+
+export const BUCKET_NAME_MIN_LENGTH_LIMIT = 3;
+
+export const DESCRIPTION_LENGTH_LIMIT = 256;
